+++ packages/orq-rc/src/models/operations/updateidentity.ts
@@ -167,7 +167,7 @@ export const UpdateIdentityResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-05-09T13:10:43.586Z",
+    "2026-05-10T11:38:58.929Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {