+++ packages/orq-rc/RELEASES.md
@@ -8989,3 +8989,13 @@ Based on:
 - [typescript v4.9.0-rc.37] packages/orq-rc
 ### Releases
 - [NPM v4.9.0-rc.37] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.37 - packages/orq-rc
+
+## 2026-05-10 11:42:09
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.762.0 (2.882.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.9.0-rc.38] packages/orq-rc
+### Releases
+- [NPM v4.9.0-rc.38] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.38 - packages/orq-rc